Adjust stepping and CPU map to eschew rewiring.
Very simple really, GRBL does its step timing calculations and scheduling on one set of bitmasks,
therefore in the original setup, all GPIO pins for each operation must be on the same port.

However, it is simple enough to create the bitmask as 32-bits wide, let GRBL do all its math,
and then split it in half and write it to two separate ports as necessary.